requirements.txt
anyio==3.6.2
asttokens==2.2.1
attrs==23.1.0
Automat==22.10.0
backcall==0.2.0
beanie==1.19.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
constantly==15.1.0
construct==2.10.68
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.95.1
filelock==3.12.0
flatbuffers==23.5.9
fonttools==4.39.4
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
humanfriendly==10.0
hyperlink==21.0.0
idna==3.4
importlib-resources==5.12.0
incremental==22.10.0
iniconfig==2.0.0
ipython==8.13.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
kiwisolver==1.4.4
lazy-model==0.0.5
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
motor==3.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
onnx==1.14.0
onnxruntime==1.14.1
opencv-python==4.7.0.72
orjson==3.8.12
packaging==23.1
pandas==2.0.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
pluggy==1.0.0
ply==3.11
prompt-toolkit==3.0.38
protobuf==4.23.0
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pycryptodomex==3.17
pydantic==1.10.7
Pygments==2.15.1
pymongo==4.3.3
pyparsing==3.0.9
pyreadline3==3.4.1
pysmi==0.3.4
pysnmp==4.4.12
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
requests==2.30.0
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.22.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.26.1
sympy==1.12
tensor==0.3.6
thop==0.1.1.post2209072238
toml==0.10.2
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
traitlets==5.9.0
Twisted==22.10.0
twisted-iocpsupport==1.0.3
typing_extensions==4.5.0
tzdata==2023.3
ujson==5.7.0
ultralytics==8.0.99
urllib3==1.26.15
uvicorn==0.22.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
zipp==3.15.0
zope.interface==6.0
